# open-commerce-search-api/src/main/resources/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Open Commerce Search API",
    "description" : "A common product search API that separates its usage from required search expertise",
    "contact" : {
      "email" : "info@commerce-experts.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "0.2"
  },
  "paths" : {
    "/indexer-api/v1/full/cancel" : {
      "post" : {
        "description" : "Cancels the import and in case there was an index created, it will be deleted.",
        "operationId" : "cancel",
        "requestBody" : {
          "$ref" : "#/components/requestBodies/ImportSession"
        },
        "responses" : {
          "400" : {
            "description" : "indexation was already confirmed or import session is invalid"
          }
        },
        "servers" : [ {
          "url" : "http://indexer-service",
          "description" : "Service to run a full import into a new index. To do so, start a indexation session with a request to 'start' and use the returned ImportSession object to 'add' products bulkwise. If all documents where added, use the 'done' request to deploy that index. In case there were failures (or more failures then tollerated), the 'cancel' request can be used to stop the process and cleanup the incomplete index. Depending on the document size, an amount of 500-2000 documents per bulk is sufficient. If product and content data should be indexed, its recommended to use different indexes. Otherwise make sure to use the same fields for the same content type, e.g. both kind of documents can have a textual 'title' field, but both kind of documents shouldn't have for example an 'author' field, which could be used for product facets (e.g. book authors) but not for faceting the content documents (e.g. blog post authors).",
          "variables" : { }
        } ]
      }
    },
    "/indexer-api/v1/full/done" : {
      "post" : {
        "description" : "Finishes the import, flushing the new index and (in case there is already an index with the initialized name) replacing the old one.",
        "operationId" : "done",
        "requestBody" : {
          "$ref" : "#/components/requestBodies/ImportSession"
        },
        "responses" : {
          "200" : {
            "description" : "successfully done"
          },
          "400" : {
            "description" : "indexation was already confirmed or import session is invalid"
          }
        },
        "servers" : [ {
          "url" : "http://indexer-service",
          "description" : "Service to run a full import into a new index. To do so, start a indexation session with a request to 'start' and use the returned ImportSession object to 'add' products bulkwise. If all documents where added, use the 'done' request to deploy that index. In case there were failures (or more failures then tollerated), the 'cancel' request can be used to stop the process and cleanup the incomplete index. Depending on the document size, an amount of 500-2000 documents per bulk is sufficient. If product and content data should be indexed, its recommended to use different indexes. Otherwise make sure to use the same fields for the same content type, e.g. both kind of documents can have a textual 'title' field, but both kind of documents shouldn't have for example an 'author' field, which could be used for product facets (e.g. book authors) but not for faceting the content documents (e.g. blog post authors).",
          "variables" : { }
        } ]
      }
    },
    "/indexer-api/v1/full/start/{indexName}" : {
      "get" : {
        "summary" : "Starts a new full import",
        "description" : "Starts a new full import. Returns a handle containing meta data, that has to be passed to all following calls.",
        "operationId" : "startImport",
        "parameters" : [ {
          "name" : "indexName",
          "in" : "path",
          "description" : "index name, that should match the regular expression '[a-z0-9_-]+'",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "locale",
          "in" : "query",
          "description" : "used for language dependent settings",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "type" : "string"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "$ref" : "#/components/responses/ImportSession"
          },
          "409" : {
            "description" : "there is already an import running for that index"
          }
        },
        "servers" : [ {
          "url" : "http://indexer-service",
          "description" : "Service to run a full import into a new index. To do so, start a indexation session with a request to 'start' and use the returned ImportSession object to 'add' products bulkwise. If all documents where added, use the 'done' request to deploy that index. In case there were failures (or more failures then tollerated), the 'cancel' request can be used to stop the process and cleanup the incomplete index. Depending on the document size, an amount of 500-2000 documents per bulk is sufficient. If product and content data should be indexed, its recommended to use different indexes. Otherwise make sure to use the same fields for the same content type, e.g. both kind of documents can have a textual 'title' field, but both kind of documents shouldn't have for example an 'author' field, which could be used for product facets (e.g. book authors) but not for faceting the content documents (e.g. blog post authors).",
          "variables" : { }
        } ]
      }
    },
    "/indexer-api/v1/full/add" : {
      "post" : {
        "summary" : "Add documents to a running import session",
        "description" : "Add one or more documents to a running import session.",
        "operationId" : "add",
        "requestBody" : {
          "$ref" : "#/components/requestBodies/BulkImportData"
        },
        "responses" : {
          "200" : {
            "description" : "documents successfully added"
          },
          "400" : {
            "description" : "import session is invalid"
          }
        },
        "servers" : [ {
          "url" : "http://indexer-service",
          "description" : "Service to run a full import into a new index. To do so, start a indexation session with a request to 'start' and use the returned ImportSession object to 'add' products bulkwise. If all documents where added, use the 'done' request to deploy that index. In case there were failures (or more failures then tollerated), the 'cancel' request can be used to stop the process and cleanup the incomplete index. Depending on the document size, an amount of 500-2000 documents per bulk is sufficient. If product and content data should be indexed, its recommended to use different indexes. Otherwise make sure to use the same fields for the same content type, e.g. both kind of documents can have a textual 'title' field, but both kind of documents shouldn't have for example an 'author' field, which could be used for product facets (e.g. book authors) but not for faceting the content documents (e.g. blog post authors).",
          "variables" : { }
        } ]
      }
    },
...
  },
...
}